+++ fbw-a32nx/src/systems/fmgc/src/guidance/vnav/descent/DescentGuidance.ts
@@ -1,5 +1,5 @@
 // @ts-strict-ignore
-// Copyright (c) 2021-2023 FlyByWire Simulations
+// Copyright (c) 2021-2025 FlyByWire Simulations
 //
 // SPDX-License-Identifier: GPL-3.0
 
@@ -7,7 +7,7 @@ import { RequestedVerticalMode, TargetAltitude, TargetVerticalSpeed } from '@fmg
 import { GuidanceController } from '@fmgc/guidance/GuidanceController';
 import { AtmosphericConditions } from '@fmgc/guidance/vnav/AtmosphericConditions';
 import { AircraftToDescentProfileRelation } from '@fmgc/guidance/vnav/descent/AircraftToProfileRelation';
-import { NavGeometryProfile } from '@fmgc/guidance/vnav/profile/NavGeometryProfile';
+import { NavGeometryProfile, VerticalCheckpointReason } from '@fmgc/guidance/vnav/profile/NavGeometryProfile';
 import { VerticalProfileComputationParametersObserver } from '@fmgc/guidance/vnav/VerticalProfileComputationParameters';
 import { Arinc429Word } from '@flybywiresim/fbw-sdk';
 import { VerticalMode } from '@shared/autopilot';
@@ -218,12 +218,27 @@ export class DescentGuidance {
     const isSpeedAuto = Simplane.getAutoPilotAirspeedManaged();
     const isApproachPhaseActive = this.observer.get().flightPhase === FmgcFlightPhase.Approach;
     const isHoldActive = this.guidanceController.isManualHoldActive();
+    const isCruisePhase = this.observer.get().flightPhase === FmgcFlightPhase.Cruise;
+    const hasStepDescent = this.aircraftToDescentProfileRelation.profile?.checkpoints.some(
+      (checkpoint) => checkpoint.reason === VerticalCheckpointReason.StepDescent,
+    );
     const targetVerticalSpeed = this.aircraftToDescentProfileRelation.currentTargetVerticalSpeed();
 
     this.targetAltitudeGuidance = this.atmosphericConditions.currentPressureAltitude - linearDeviation;
 
     this.updatePathCaptureState(linearDeviation, targetVerticalSpeed);
     const shouldGoOffPath = this.pathCaptureState === PathCaptureState.OffPath;
+
+    if (isCruisePhase) {
+      this.requestedVerticalMode = RequestedVerticalMode.VsSpeed;
+      this.targetVerticalSpeed = -1000;
+      return;
+    }
+
+    if (hasStepDescent && this.observer.get().flightPhase === FmgcFlightPhase.Cruise) {
+      this.changeState(DescentVerticalGuidanceState.Observing);
+      return;
+    }
 
     if ((!isHoldActive && shouldGoOffPath && linearDeviation > 0) || this.isInOverspeedCondition) {
       // above path
@@ -387,6 +402,7 @@ export class DescentGuidance {
   private updateSpeedMarginState() {
     const { flightPhase } = this.observer.get();
     const isHoldActive = this.guidanceController.isManualHoldActive();
+    const isStepDescentActive = VerticalCheckpointReason.StepDescent;
 
     if (flightPhase !== FmgcFlightPhase.Descent) {
       this.changeSpeedState(DescentSpeedGuidanceState.NotInDescentPhase);
@@ -395,6 +411,7 @@ export class DescentGuidance {
 
     const shouldShowMargins =
       !isHoldActive &&
+      !isStepDescentActive &&
       this.verticalState === DescentVerticalGuidanceState.ProvidingGuidance &&
       Simplane.getAutoPilotAirspeedManaged();
     this.changeSpeedState(