+++ .github/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 1. [A32NX/SFCC] Added implementation of SAP and FAP in SFCC - @Eagle941 (Joe)
 1. [A32NX/SFCC] Connected SFCC to electrical busses with simulation of power consumption and power restoration behaviour - @Eagle941 (Joe)
+1. [A32NX/FMS] Limit step descent guidance to -1000fpm - @Jonny23787 (Jonathan)
 
 ## 0.14.0